+++ Runtime/Module/H2D/H2DJumpModule.cs
@@ -10,19 +10,21 @@ namespace Motivation
     [CreateAssetMenu(fileName = "H2DJumpModule", menuName = "Motivation/H2D/Jump", order = 2)]
     public class H2DJumpModule : ControllerModule
     {
-        [Header("配置")]
-        [SerializeField, Title("跳跃向上速度")] private int speed = 5;
-        [SerializeField, Title("最大跳跃次数")] private int maxJumpCnt = 2;
-        [SerializeField, Title("按键缓冲时间")] private float jumpBufferTime = 0.1f;
-        [SerializeField, Title("土狼时间")] private float coyotoTime = 0.1f;
-
         [Header("键位")]
-        [SerializeField] private List<KeyCode> jumpKeys = new List<KeyCode> { KeyCode.Space, KeyCode.W };
+        [SerializeField] protected List<KeyCode> jumpKeys = new List<KeyCode> { KeyCode.Space, KeyCode.W };
+
+        [Header("配置")]
+        [SerializeField, Title("跳跃向上速度")] protected int speed = 5;
+        [SerializeField, Title("最大跳跃次数")] protected int maxJumpCnt = 2;
+        [SerializeField, Title("按键缓冲时间")] protected float jumpBufferTime = 0.1f;
+        [SerializeField, Title("土狼时间")] protected float coyotoTime = 0.1f;
 
         public int MaxJumpCount { get => maxJumpCnt; set => maxJumpCnt = value; }
-        private int jumpCnt = 0;
+        protected int jumpCnt = 0;
 
-        private DelayedTrigger jumpBuffer, groundBuffer;
+        protected virtual uint GroundStateMask { get; } = MotivatorState.Grounded;
+
+        protected DelayedTrigger jumpBuffer, groundBuffer;
 
         public override void OnAdd(Motivator m)
         {
@@ -31,14 +33,14 @@ namespace Motivation
 
             jumpBuffer = new DelayedTrigger(jumpBufferTime);
             groundBuffer = new DelayedTrigger(coyotoTime);
-            if (Host.MatchAny(MotivatorState.Grounded)) groundBuffer.Trigger();
+            if (Host.MatchAny(GroundStateMask)) groundBuffer.Trigger();
 
             Host.RegisterKeys(jumpKeys.ToArray());
         }
 
         public override void OnStateChange(uint state)
         {
-            if (Host.MatchAny(MotivatorState.Grounded))
+            if (Host.MatchAny(GroundStateMask))
             {
                 jumpCnt = 0;
                 groundBuffer.Trigger();
@@ -49,7 +51,7 @@ namespace Motivation
         {
             jumpBuffer.Tick(time);
 
-            if (!Host.MatchAny(MotivatorState.Grounded))
+            if (!Host.MatchAny(GroundStateMask))
                 groundBuffer.Tick(time);
         }